removing specifiers question tests from allinone
also: ignore unused declared dependency warnings for auto-service in bcp
